Adjust attributes usage
