+++ source/backtraceApi.ts
@@ -5,31 +5,26 @@ import { BacktraceResult } from './model/backtraceResult';
 
 export class BacktraceApi {
   private axiosInstance: AxiosInstance;
-  
-  constructor(
-    private backtraceUri: string,
-    private timeout: number,
-    private ignoreSslCert: boolean,
-  ) {
-    this.axiosInstance = axios.create();
+
+  constructor(backtraceUri: string, timeout: number, ignoreSslCert: boolean) {
+    this.axiosInstance = axios.create({
+      baseURL: backtraceUri,
+      timeout,
+      httpsAgent: ignoreSslCert
+        ? new Agent({
+            rejectUnauthorized: false,
+          })
+        : globalAgent,
+      headers: {
+        'Content-Type': 'undefined', // https://stackoverflow.com/questions/39280438/fetch-missing-boundary-in-multipart-form-data-post
+      },
+    });
   }
 
   public async send(report: BacktraceReport): Promise<BacktraceResult> {
     try {
       const formData = await report.toFormData();
-      const options = {
-          baseURL: this.backtraceUri,
-          timeout: this.timeout,
-          httpsAgent: this.ignoreSslCert
-            ? new Agent({
-                rejectUnauthorized: false,
-              })
-            : globalAgent,
-          headers: {
-            'Content-Type': 'undefined', // https://stackoverflow.com/questions/39280438/fetch-missing-boundary-in-multipart-form-data-post
-          },
-      };
-      const result = await this.axiosInstance.post(this.backtraceUri, formData, options);
+      const result = await this.axiosInstance.post('', formData);
 
       if (result.status === 429) {
         const err = new Error(`Backtrace - reached report limit.`);
