+++ source/backtraceClient.ts
@@ -1,15 +1,18 @@
 import { BacktraceApi } from './backtraceApi';
 import { ClientRateLimit } from './clientRateLimit';
+import { pageStartTime } from './index';
 import { BacktraceClientOptions } from './model/backtraceClientOptions';
 import { BacktraceReport } from './model/backtraceReport';
 import { BacktraceResult } from './model/backtraceResult';
 import { Breadcrumbs } from './model/breadcrumbs';
+import { getBrowserName, getBrowserVersion, getOs, isMobile } from './utils/agentUtils';
 /**
  * Backtrace client
  */
 export class BacktraceClient {
   public options: BacktraceClientOptions;
   public breadcrumbs: Breadcrumbs;
+  public readonly attributes: { [index: string]: any };
 
   private _backtraceApi: BacktraceApi;
   private _clientRateLimit: ClientRateLimit;
@@ -26,6 +29,10 @@ export class BacktraceClient {
     this._backtraceApi = new BacktraceApi(this.getSubmitUrl(), this.options.timeout, this.options.ignoreSslCert);
     this._clientRateLimit = new ClientRateLimit(this.options.rateLimit);
     this.registerHandlers();
+    this.attributes = {
+      ...this.readAttributes(),
+      ...this.options.userAttributes,
+    };
   }
 
   /**
@@ -38,17 +45,15 @@ export class BacktraceClient {
   public memorize(key: string, value: any): void {
     (this.options.userAttributes as any)[key] = value;
   }
-  
+
   public createReport(
-      payload: Error | string,
-      reportAttributes: object | undefined = {}, 
-      attachment?: string | object,
+    payload: Error | string,
+    reportAttributes: object | undefined = {},
+    attachment?: string | object,
   ): BacktraceReport {
-    // this.emit('new-report', payload, reportAttributes);
-    const attributes = this.combineClientAttributes(reportAttributes);
     const breadcrumbs = this.breadcrumbs.isEnabled() ? this.breadcrumbs.get() : undefined;
-    const report = new BacktraceReport(payload, attributes, breadcrumbs, attachment);
-    
+    const report = new BacktraceReport(payload, reportAttributes, breadcrumbs, attachment);
+
     report.send = (callback) => {
       this.sendAsync(report)
         .then(() => {
@@ -100,9 +105,9 @@ export class BacktraceClient {
    * @param attachment data to be converted to a Blob and sent as attachment with report
    */
   public reportSync(
-      payload: Error | string,
-      reportAttributes: object | undefined = {}, 
-      attachment?: string | object
+    payload: Error | string,
+    reportAttributes: object | undefined = {},
+    attachment?: string | object,
   ): BacktraceResult {
     const report = this.createReport(payload, reportAttributes, attachment);
     return this.sendReport(report);
@@ -119,6 +124,10 @@ export class BacktraceClient {
     if (limitResult) {
       return limitResult;
     }
+
+    // apply default attributes
+    report.addObjectAttributes(this.attributes);
+
     this._backtraceApi
       .send(report)
       .then((result) => {
@@ -126,15 +135,15 @@ export class BacktraceClient {
           this.breadcrumbs.add(
             'Report sent to Backtrace',
             {
-              error: result.Error, 
-              message: result.Message, 
-              objectId: result.ObjectId,  
-            }, 
+              error: result.Error,
+              message: result.Message,
+              objectId: result.ObjectId,
+            },
             Date.now(),
-            'error', 
-            'log'
+            'error',
+            'log',
           );
-        };
+        }
 
         if (callback) {
           callback(result.Error, result);
@@ -232,6 +241,44 @@ export class BacktraceClient {
         'exception.lineNumber': lineNumber,
         'exception.columnNumber': columnNumber,
       });
+    };
+  }
+
+  private readAttributes(): { [index: string]: any } {
+    const browserName = getBrowserName();
+    return {
+      application: document.title,
+      'process.age': Math.floor((new Date().getTime() - pageStartTime.getTime()) / 1000),
+      hostname: window.location && window.location.hostname,
+      referer: window.location && window.location.href,
+      'user.agent.full': navigator.userAgent,
+      'location.port': document.location.port,
+      'location.protocol': document.location.protocol,
+      'location.origin': window.location.origin,
+      'location.href': window.location.href || document.URL,
+      language: navigator.language,
+      'browser.name': browserName,
+      'browser.version': getBrowserVersion(browserName),
+      'browser.platform': navigator.platform,
+      'browser.vendor': navigator.vendor,
+      'cookies.enable': navigator.cookieEnabled,
+      'document.domain': document.domain,
+      'document.baseURI': document.baseURI,
+      'document.title': document.title,
+      'document.referrer': document.referrer,
+      mobile: isMobile(),
+      'localstorage.enable': !!window.localStorage,
+      'uname.sysname': getOs(),
+      'window.innerHeight': window.innerHeight,
+      'window.innerWidth': window.innerWidth,
+      'window.outerHeight': window.outerHeight,
+      'window.outerWidth': window.outerWidth,
+      'window.pageXOffset': window.pageXOffset,
+      'window.pageYOffset': window.pageYOffset,
+      'window.screenX': window.screenX,
+      'window.screenY': window.screenY,
+      'window.screenLeft': window.screenLeft,
+      'window.screenTop': window.screenTop,
     };
   }
 }
